Search 추가
- 리뷰 평점 추가
- total count 추가

+++ src/main/java/com/alconn/copang/item/dto/ItemDetailForm.java
@@ -64,7 +64,7 @@ public class ItemDetailForm {
 
         private Long sellerId;
 
-        private Integer averageRating;
+        private Double averageRating;
 
         private ShipmentInfoForm shipmentInfoForm;
     }
